Move Comparable, define one_left, one_right
The methods to compare node values have been moved to their own module:
Comparable. This module is saved in `comparable.rb`

Wrote basic methods to find indices of the root node of a sorted array,
as well as the indices one to the left and one to the right. Plan to use
these for arbitrary nodes, not just `root